seperated make and make_checks
