+++ wiggle_query_language/clauses/make/README.md
@@ -1,1 +1,11 @@
 # Make
+
+The MAKE clause in the
+
+## Introduction
+
+
+### Aims
+
+
+## Examples
