+++ wiggle_query_language/clauses/make/make.py
@@ -1,83 +1,15 @@
 from typing import Optional
 
-from exceptions.wql.make import (
-    MakeClauseSyntaxError,
-    MakeParamSyntaxError,
-    MakeNonDirectedRelationshipError,
-    MakeIllegalCharacter,
-)
 from models.wql.parsed_query import ParsedMake
 from wiggle_query_language.clauses.regexes.make import (
     MAKE_STATEMENT_ALL_REGEX,
-    MAKE_STATEMENT_CHECK_CLAUSE_SYNTAX,
-    MAKE_STATEMENT_CHECK_PARAMS_SYNTAX,
     NODES_RELS_PATTERN_REGEX,
-    RELATIONSHIP_DIR_CHECK_REGEX,
-    PARAM_LIST_VALUE,
-    ILLEGAL_CHARS_REGEX,
 )
 
-
-def check_param_formatting(params_string: str) -> bool:
-    """
-    Chesk the
-    :param params_string:
-    :return:
-    """
-
-    exp_param_count = params_string.count(",") + 1
-    colon_count = params_string.count(":")
-    if exp_param_count != colon_count:
-        raise MakeParamSyntaxError(f"SyntaxError: {params_string} missing : or ,")
-
-    return True
-
-
-def check_make_params(make_matches: list[str]) -> True:
-    """
-    Very crude check that the params match up with the colons
-    :param make_matches:  The extracted MAKE statements.
-    :return: A bool for testing.
-    """
-
-    for stmt in make_matches:
-        if not (param_string := MAKE_STATEMENT_CHECK_PARAMS_SYNTAX.findall(stmt)):
-            continue
-
-        # TODO remove double loop, most of the time will be one match..
-        for param_match in param_string:
-            # remove the list from the params
-            params_sans_list = PARAM_LIST_VALUE.sub("", param_match)
-            check_param_formatting(params_sans_list)
-
-            # check list
-            if params_lists := PARAM_LIST_VALUE.findall(param_match):
-                for params_list in params_lists:
-                    try:
-                        # TODO remove this..
-                        eval(params_list)
-                    except SyntaxError:
-                        raise MakeParamSyntaxError(
-                            f"SyntaxError: {params_lists} missing a comma?"
-                        )
-
-    return True
-
-
-def check_make_clause_syntax(query_string: str) -> None:
-    """
-    Checks the syntax of the MAKE statement.
-    :param query_string: The extracted MAKE statements.
-    :return: None or an Exception
-    """
-
-    if matches := MAKE_STATEMENT_CHECK_CLAUSE_SYNTAX.findall(query_string):
-        for match in matches:
-            raise MakeClauseSyntaxError(
-                f"SyntaxError: {match} was not recognised did you mean MAKE?"
-            )
-
-    return None
+from wiggle_query_language.clauses.make.make_checks import (
+    validate_make_statement,
+    check_make_clause_syntax,
+)
 
 
 def extract_all_make_statements(query_string: str) -> Optional[list[str]]:
@@ -112,52 +44,6 @@ def build_parsed_make(statement: str) -> ParsedMake:
     )
 
     return parsed_make
-
-
-def check_relationships(make_matches: list[str]) -> bool:
-    """
-    Checks to see that the created relationship is directed.
-    :param make_matches: The extracted make statements.
-    :return: A bool for testing.
-    """
-
-    for stmt in make_matches:
-        rels = RELATIONSHIP_DIR_CHECK_REGEX.findall(stmt)
-
-        for rel in rels:
-            if "<" in rel and ">" in rel:
-                raise MakeNonDirectedRelationshipError(
-                    "Relationships must be unidirectional"
-                )
-            if "<" not in rel and ">" not in rel:
-                raise MakeNonDirectedRelationshipError(
-                    "Relationships must be singly directed"
-                )
-
-            continue
-
-    return True
-
-
-def check_illegal_characters(make_matches: list[str]) -> bool:
-    for stmt in make_matches:
-        if match := ILLEGAL_CHARS_REGEX.search(stmt):
-            raise MakeIllegalCharacter(f"{match.group()} is not allowed")
-
-    return True
-
-
-def validate_make_statement(make_matches: list[str]) -> bool:
-    """
-    Handles the validation for the make statement
-    :param make_matches: The extracted make statements
-    :return: The
-    """
-    check_illegal_characters(make_matches)
-    check_make_params(make_matches)
-    check_relationships(make_matches)
-
-    return True
 
 
 def parse_make_statement_from_query_string(
